move encode/decode into common/utils
