+++ common/utils.go
@@ -366,3 +366,40 @@ func IsValidSortField(sort string, allowedFields []string) bool {
 	}
 	return false
 }
+
+// EncodeEncryptMessage encrypts and encodes a message using AES-GCM
+func EncodeEncryptMessage(message []byte, hexEncryptionKey string) (string, error) {
+	// First base64 encode the message
+	base64EncodedMessage := base64.StdEncoding.EncodeToString(message)
+
+	// Then encrypt using AES-GCM
+	encryptedMessage, err := EncryptGCM(base64EncodedMessage, hexEncryptionKey)
+	if err != nil {
+		return "", fmt.Errorf("failed to encrypt message: %w", err)
+	}
+
+	return encryptedMessage, nil
+}
+
+// DecodeDecryptMessage decodes and decrypts a message using AES-GCM
+func DecodeDecryptMessage(encodedMessage, hexEncryptionKey string) ([]byte, error) {
+	// First decode from base64
+	encryptedMessage, err := base64.StdEncoding.DecodeString(encodedMessage)
+	if err != nil {
+		return nil, fmt.Errorf("failed to decode base64: %w", err)
+	}
+
+	// Decrypt using AES-GCM
+	decryptedMessage, err := DecryptGCM(encryptedMessage, hexEncryptionKey)
+	if err != nil {
+		return nil, fmt.Errorf("failed to decrypt message: %w", err)
+	}
+
+	// The decrypted message is a base64-encoded string, so decode it
+	inboundBody, err := base64.StdEncoding.DecodeString(string(decryptedMessage))
+	if err != nil {
+		return nil, fmt.Errorf("failed to decode inbound message: %w", err)
+	}
+
+	return inboundBody, nil
+}
