+++ internal/protocol/reshare.go
@@ -89,7 +89,7 @@ func CreateQcSetupAndReshare(v *vaultType.Vault, sessionID, encryptionKey, local
 
 	// Encrypt and upload setup message
 	relayClient := relay.NewRelayClient("https://api.vultisig.com/router")
-	encryptedSetup, err := encodeEncryptMessage(setupMsg, encryptionKey)
+	encryptedSetup, err := common.EncodeEncryptMessage(setupMsg, encryptionKey)
 	if err != nil {
 		return fmt.Errorf("failed to encrypt setup message: %w", err)
 	}
@@ -257,7 +257,7 @@ func ProcessQcReshareInbound(mpcWrapper *vault.MPCWrapperImp, handle vault.Handl
 				}
 
 				// Decrypt the message using the same method as vault service
-				inboundBody, err := decodeDecryptMessage(message.Body, encryptionKey)
+				inboundBody, err := common.DecodeDecryptMessage(message.Body, encryptionKey)
 				if err != nil {
 					continue
 				}
@@ -410,41 +410,4 @@ func getCommitteeIndices(allCommittee, oldCommittee, newCommittee []string) ([]i
 	}
 
 	return oldIndices, newIndices
-}
-
-// encodeEncryptMessage encrypts and encodes a message using AES-GCM
-func encodeEncryptMessage(message []byte, hexEncryptionKey string) (string, error) {
-	// First base64 encode the message
-	base64EncodedMessage := base64.StdEncoding.EncodeToString(message)
-
-	// Then encrypt using AES-GCM
-	encryptedMessage, err := common.EncryptGCM(base64EncodedMessage, hexEncryptionKey)
-	if err != nil {
-		return "", fmt.Errorf("failed to encrypt message: %w", err)
-	}
-
-	return encryptedMessage, nil
-}
-
-// decodeDecryptMessage decodes and decrypts a message using AES-GCM
-func decodeDecryptMessage(encodedMessage, hexEncryptionKey string) ([]byte, error) {
-	// First decode from base64
-	encryptedMessage, err := base64.StdEncoding.DecodeString(encodedMessage)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode base64: %w", err)
-	}
-
-	// Decrypt using AES-GCM
-	decryptedMessage, err := common.DecryptGCM(encryptedMessage, hexEncryptionKey)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decrypt message: %w", err)
-	}
-
-	// The decrypted message is a base64-encoded string, so decode it
-	inboundBody, err := base64.StdEncoding.DecodeString(string(decryptedMessage))
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode inbound message: %w", err)
-	}
-
-	return inboundBody, nil
 }
